requirements.txt
alembic==1.5.8
apache-airflow==2.0.2
apache-airflow-providers-ftp==1.0.1
apache-airflow-providers-http==1.1.1
apache-airflow-providers-imap==1.0.1
apache-airflow-providers-sqlite==1.0.2
apispec==3.3.2
appdirs==1.4.4
argcomplete==1.12.3
attrs==20.3.0
autopep8==1.5.6
Babel==2.9.1
beautifulsoup4==4.9.3
black==20.8b1
blinker==1.4
bs4==0.0.1
cached-property==1.5.2
cattrs==1.6.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
colorlog==5.0.1
commonmark==0.9.1
connexion==2.7.0
croniter==0.3.37
crypto==1.4.1
cryptography==3.4.7
cssselect==1.1.0
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.3
distlib==0.3.1
dnspython==2.1.0
docutils==0.17.1
email-validator==1.1.2
filelock==3.0.12
fire==0.4.0
Flask==1.1.2
Flask-AppBuilder==3.2.3
Flask-Babel==1.0.0
Flask-Caching==1.10.1
Flask-Cors==3.0.10
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
gcloud==0.18.3
gdown==3.12.2
googleapis-common-protos==1.53.0
googledrivedownloader==0.4
graphviz==0.16
gunicorn==19.10.0
httplib2==0.19.1
idna==2.10
importlib-metadata==1.7.0
importlib-resources==1.5.0
inflection==0.5.1
ipython-genutils==0.2.0
iso8601==0.1.14
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-core==4.3.0
jwcrypto==1.0
lazy-object-proxy==1.6.0
lockfile==0.12.2
lxml==4.6.3
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.11.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==2.1.0
marshmallow-sqlalchemy==0.23.1
MouseInfo==0.1.3
mypy-extensions==0.4.3
Naked==0.1.31
natsort==7.1.1
numpy==1.20.2
oauth2client==4.1.3
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.0
opencv-python==4.5.1.48
pandas==1.2.4
parse==1.19.0
parso==0.8.2
pathspec==0.8.1
pendulum==2.1.2
pickleshare==0.7.5
Pillow==8.2.0
pipenv==2020.11.15
plyer==2.0.0
prison==0.1.3
prompt-toolkit==1.0.15
protobuf==3.18.0
psutil==5.8.0
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.10.4
PyGetWindow==0.0.9
Pygments==2.8.1
PyJWT==1.7.1
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
pyquery==1.4.3
Pyrebase4==4.5.0
PyRect==0.1.4
pyrsistent==0.17.3
PyScreeze==0.1.26
PySimpleGUI==4.39.1
PySocks==1.7.1
python-daemon==2.3.0
python-dateutil==2.8.1
python-editor==1.0.4
python-jwt==3.3.0
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytube @ git+https://github.com/ssuwani/pytube@79985cde70fe78a15c36ca6c732ed4558a6902ec
PyTweening==1.0.3
pytz==2021.1
pytzdata==2020.1
pywebcopy==6.3.0
PyYAML==5.4.1
pyzmq==22.1.0
regex==2021.3.17
requests==2.25.1
requests-toolbelt==0.9.1
rich==9.2.0
rsa==4.7.2
scapy==2.4.5
selenium==3.141.0
setproctitle==1.2.2
shellescape==3.8.1
simplegeneric==0.8.1
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.37.0
swagger-ui-bundle==0.0.8
tabulate==0.8.9
tenacity==6.2.0
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.60.0
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions==3.7.4.3
unicodecsv==0.14.1
urllib3==1.26.4
virtualenv==20.4.4
virtualenv-clone==0.5.4
w3lib==1.22.0
wcwidth==0.2.5
Werkzeug==1.0.1
wget==3.2
wrapt==1.12.1
WTForms==2.3.3
zipp==3.4.1
